requirements.txt
gymnasium>=0.29.0
numpy>=1.24.0
networkx>=3.0

stable-baselines3>=2.1.0
tensorboard>=2.14.0


torch>=2.0.0
torch-geometric>=2.4.0


catboost>=1.2

matplotlib>=3.7.0

scipy>=1.11.0

pyyaml>=6.0

pytest>=7.0.0
